requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
supabase==2.3.0
asyncpg==0.29.0

# AI
anthropic==0.25.0

# Integrations (versão compatível com supabase)
httpx==0.24.1
aiohttp==3.9.1

# MegaAPI (WhatsApp)
requests==2.31.0

# Email
aiosmtplib==3.0.1
email-validator==2.1.0